refactor: handling book not found when deleting book

+++ src/main/java/com/box/library/book/BookService.java
@@ -27,6 +27,9 @@ public class BookService {
     }
 
     public void deleteById(Long bookId) {
+        if(!repository.existsById(bookId)) {
+            throw new BookNotFoundException(bookId);
+        }
         repository.deleteById(bookId);
     }
 }
